Added a working version of the game
This is a working version of the Flappy Bird Microbit game. It has scrolling walls and keeps score. In the future I would like to control the bird by tilting the Microbit and using accelerometer information rather than by pressing buttons. I may also like to add a replay the game function controlled by shaking the Microbit instead of requiring the reset button to be pressed. It might also be fun to add a sound track to the game; I am not sure if it would become annoying or interfere with the score sounds.